+++ src/app/pages/data/data.module.ts
@@ -1,11 +1,8 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
 
 import { ConfirmationService, DataTableModule, DialogModule, ButtonModule, GrowlModule, MessagesModule, InputTextModule,
     ConfirmDialogModule, CalendarModule, DropdownModule, FileUploadModule, CheckboxModule, TooltipModule }
     from 'primeng/primeng';
-
-import { BiosysCoreModule } from '../../biosys-core/biosys-core.module';
 
 import { SharedModule } from '../../shared/shared.module';
 
@@ -16,7 +13,7 @@ import { EditRecordComponent } from './edit-record/edit-record.component';
 
 
 @NgModule({
-    imports: [CommonModule, SharedModule, BiosysCoreModule, DataTableModule, DialogModule, ButtonModule, GrowlModule,
+    imports: [SharedModule, DataTableModule, DialogModule, ButtonModule, GrowlModule,
         MessagesModule, ConfirmDialogModule, CalendarModule, DropdownModule, FileUploadModule, CheckboxModule,
         InputTextModule, TooltipModule],
     declarations: [DataListProjectsComponent, ListDatasetsComponent, ManageDataComponent, EditRecordComponent],
